+++ docs/source/NEWS.rst
@@ -1,13 +1,17 @@
 Changelog
 =========
 
-Release 24.0.0 (2024-09-16)
+Release 24.0.0 (2024-10-01)
 ---------------------------
 
 API Changes
 ^^^^^^^^
 
 - This is the last release that supports Python <3.8 and MongoDB <4.0
+- Methods `insert()`, `update()` and `remove()` are deprecated in favor of corresponding `*_one()`
+  and `*_many()` methods. Old methods will be removed in the next release.
+- `as_class` argument of `find()`, `find_with_cursor()` and `find_one()` is deprecated and will
+  be removed in the next release.
 
 
 Release UPCOMING (yyyy-mm-dd)